requirements.txt
Django==4.0.6
djangorestframework==3.13.1
httplib2==0.20.4
requests==2.28.1
pillow